+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+3.6.0 - January, 07 2021
+
+* 9f59986 Feat: setInclude | warn on already excluded values (#121) (Jan Peer Stöcklmair)
+* 2907583 Chore(deps): bump lodash from 4.17.15 to 4.17.20 (#118) (dependabot[bot])
+
+3.5.1 - November, 12 2020
+
+* 0537718 Fix: get values even if they have some line breaks and whitespaces (JPeer264)
+* af50ba1 Docs: add Analyse.org (Jan Peer Stöcklmair)
+
+3.5.0 - September, 23 2020
+
+* 076761e Feat: replace vue class selectors (#117) (Jan Peer Stöcklmair)
+
 3.4.0 - August, 09 2020
 
 * c32aa0e Docs: improve optimize docs (JPeer264)